+++ frontend/src/pages/StudyBoardDetail/StudyBoardDetail.tsx
@@ -1,17 +1,15 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import Header from './components/Header';
-import Body from './components/Body';
-import ContentLayout from '../../components/ContentLayout';
-import Loading from '../../components/Loading/Loading';
 import { useQueries } from 'react-query';
-import { getBoardDetail, getCommentAll } from '../../apis/apiClient';
+import { Header, Body } from './components';
+import { ContentLayout, Loading } from '@components/index';
+import { getBoardDetail, getCommentAll } from '@apis/apiClient';
 
 const StudyBoardDetail = () => {
   const { id } = useParams();
   const allQuery = useQueries([
     {
-      queryKey: ['data', id],
+      queryKey: ['board', id],
       queryFn: async () => await getBoardDetail(String(id)),
     },
     {
@@ -24,7 +22,11 @@ const StudyBoardDetail = () => {
   if (loading) return <Loading />;
   return (
     <ContentLayout>
-      <Header board={allQuery[0].data} comment={allQuery[1].data} />
+      <Header
+        board={allQuery[0].data}
+        comment={allQuery[1].data}
+        refetch={allQuery[0].refetch}
+      />
       <Body
         board={allQuery[0].data}
         comment={allQuery[1].data}
